add bootstrap mods loading priority test

+++ src/Contracts/Bootable.php
@@ -11,9 +11,9 @@ namespace Carno\Console\Contracts;
 interface Bootable
 {
     /**
-     * @return int
+     * @return float
      */
-    public function priority() : int;
+    public function priority() : float;
 
     /**
      * @return bool
